+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ﻿# Changelog
 
+## Version 2.3.0
+* Added a brand new Mod Settings GUI, accessible from the menu list
+* Don't show server side configs in the GUI when a client has no admin rights
+* Added two new custom GUI containers for mods to use (in front and behind Valheim's own GUI)
+* Marked the old PixelFix container as obsolete
+
 ## Version 2.2.3
 * Don't use own canvas for custom GUI (fixes compat with VHVR for example)
 * Compile against BepInEx v5.4.15